+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 0.8.0 (2019-09-24)
 
 - **[Breaking change]** Update to `avm1-tree@0.8`.